Decouple Event Layer from Reb_Devices
The event device offered two hook points: one for "query" and one
for "connect".  Both of these were more or less abuses of the terms;
query was actually "Wait in Milliseconds", and connect was a strange
no-op that perpetually kept an inactive request alive (as a way of
showing the event port was "on?")

Removing the device layer and changing the query into a per-platform
method of waiting while being interruptible by system events like
Ctrl-C seems to keep it working, while moving it toward some amount
of comprehensibility.