Remove "sock"/REBREQ from DNS extension
The DNS extension was formulated as a PORT! because there was support
for asynchronously looking up a hostname.  Microsoft deprecated
the asynchronous host lookup APIs and did not carry them forward on
IpV6, saying that you needed to make your own thread and call the
synchronous version if that's something you wanted.

As asynchronous DNS was not available on POSIX, having special code
in Windows that was deprecated made little sense in the priority
scheme, so the complexities surrounding it were not worthwhile and
it was removed.

In order to continue progressing on removing the device layer to free
the code to be linked against a more solid foundation, this removes
the OPEN and CLOSE mechanic entirely--which was piggy-backing on the
network layer with a "REBREQ".  This removal means that the socket
layer will be the only client of REBREQ, allowing it to be simplified
to a bare minimum for being removed.