Decouple network extension from REBDEV, kill REBREQ
This commit eliminates the REBREQ concept entirely.

The networking layer was the last client of the device system.  It
relied on having a polling hook into the WAIT loop.  This polling hook
was used to advance progress on network send and receive, to deal with
delays in the CONNECT process, and to allow a listening socket (e.g.
for a web server) to be able to accept new connections.

What this does is to make each of those managed in the network
extension by a different linked list.  Then, the only thing that
"registering a device" now does is to register a polling hook.  The
network polling hook enumerates those three lists and handles them.

The goal was to change the internals without requiring any rewrite
of the client code.