Make READ-LINE return NULL w/no EOF on Escape
This makes it so that READ-LINE will return ~escape~ when the ESC key
is pressed.  It's distinct from the end-of-file intention, because
it actively means to disregard information that was previously
accumulated.

It also makes the console react to this by cycling to a new prompt.

Includes some bug fixes for the setting of the EOF flag.