Slightly improve error handling for networking
The `error` field had been removed from the standard port prematurely,
as it turns out it was actually being used.  This addresses some
issues of robustness with regards to erros that occur during socket
transfer--though the real answer is that the way ports handle events
is very brittle and needs to be completely replaced.