Kill RSM_OPEN/CONNECTED flags in favor of -1 socket ID
There were flags which was being maintained to say whether a socket
was open or connected, as a "Rebol Socket Mode" (RSM) for RSM_OPEN
and RSM_CONNECTED.

Rather than track these flags independently, this just takes advantage
of the fact that -1 is not a legal socket ID (indicates error return
from connect()).  By making that signal a disconnected socket, it's
a stronger way of making sure that one isn't missing out by not
remembering to check the flags...because trying to use the socket or
file descriptor when it's not ready will cause an error.  This makes
it more self-checking.