Add extra FEED_FLAG_NEXT_ARG_FROM_OUT check
When an enfix operation evaluates, the trick is that the previous
expression's output is in the frame's output cell...but it gets used
as the first argument of the function evaluation.

The flag that dictates this behavior is on the feed--despite the fact
that feeds have no concept of an output cell in and of themselves.  But
by putting it on the feed, it means that if a frame makes a subframe
that reuses the feed (and the output cell) then it can telegraph the
usage to that subframe.

However, it's an easy flag to overlook...so it's important to assert
that isn't happening.  This takes two steps toward making that be
checked better: if fetching is done at a feed level then the flag
should be cleared (as it has no out).  Also, if a frame is inheriting
a feed and the flag is set, the output cell must be consistent.