Add override for first item when making frame from feed
It may be desirable to create a frame out of a feed with a first
item that is not the next item in the feed (e.g. not feed->value).

The ability was added for a feature that is experimental and not
ready for use (if it ever will be), but it's a good ability to have
anyway.