Redo file and directory ports on top of libuv
This eliminates the %file-windows.c implementation, and adapts the
%file-posix.c code to use the libuv filesystem API.  Libuv is similar
to POSIX with a few twists, and they take care of the windows
details.

This particular change does not use any of the libuv callbacks--it just
passes in NULL for the callback function.  Therefore the only real
advantage so far is that it gets rid of the double implementation.
But it's a proof of concept of the usage of libuv that will be more
pertinent in the network layer...and it also means that asynchronous
file I/O is an option if the need for that comes up.

While this is kind of a complete rewrite of the ports, it doesn't
attack some issues in this first step...mostly just trying to get a
solid version of the previous presumed functionality.  This adds the
first significant instance of a test file for the filesystem, which
includes a fuzz test that writes to a BINARY! in duplicate with the
file calls and compares the output at the end.