Eliminate /ALLOW, OPEN/SEEK, and QUERY/MODE
The /SEEK refinement on OPEN was not like the other /SEEK refinements
that take an offset, but rather a "hint" that said the file would be
used in random access.  This refinement needs another name, like
/HINT, and there are other OS hints that could be put there.  The
issue is put aside for now, but /SEEK is bad so that's gone.

QUERY is being redesigned to a new extensible dialect, and so the
/MODE refinement (never implemented in R3-Alpha) can go.

/ALLOW pertained to some of the SECURE dialect that was never really
implemented, and is not a core competency of this interpreter.  So
the securing is being left to OS sandboxing/jails for now.