Transition Network code to use LibUV
This switches the TCP port to use LibUV calls.  The original port code
was not well suited for a streaming model (it seemed to act like data
came across the wire in predictable message chunks, vs. needing to
assemble messages)...and it could not have READ and WRITE calls in
flight at the same time.  But this is a step toward building a better
model that tries to keep the existing one working.

It introduces a separate CONNECT generic action, instead of having it
done by OPEN'ing an already OPEN port.  But other than that, it does
not change the HTTP or TLS, and the HTTPD server runs unchanged.