Eliminate LOOKUP Event
Windows had asynchronous DNS lookups, but POSIX did not.  It turned out
that Windows deprecated the method by which asynchronous DNS was being
done so the feature was removed.

Hence there was a residual LOOKUP event held over from that DNS.
If you opened a network port via a DNS name, however, that would still
be done in two steps.  You would OPEN the port once to do the lookup,
then once you received the lookup notification you would have to
OPEN it again which would trigger the connection.

Besides being just outright confusing, this creates a dodgy invariant
for OPEN.  The goal of implicit ports is to make it so that you can
do something like a READ or an APPEND to a FILE! and the OPEN and
CLOSE will be taken care of automatiaclly...and you get just one OPEN
and one CLOSE.  Hence OPEN must do everything for a port necessary to
get it to a usable state.

There may be some open to OPEN for "open and do not connect", but this
would not be the default.  The more granular steps of doing a lookup
and doing a connect can be available, but that's not what OPEN with
a URL! should do.