Do not change the notion of current directory on DO
When scripts are run with DO or on the command line, the caller is
potentially passing in arguments.  For arguments that represent
relative paths, those paths really only make sense in the worldview
of the caller... e.g. if they are on the command line, it would be
the relative to the shell's directory.  But a similar concept applies
for when a utility script is called by another script via DO.

In these scenarios it does not make sense to change the working
directory in the DO--because that throws away the contextual
information for interpreting arguments.  However, the historical
Rebol feature had been to change the current directory when running
a script (via filename) with DO, to be the directory the script
lived in.  This was to make it easier for that script to be able to
mention modules or resources that were relative to its location.

This commit removes that directory change for DO.  Though it keeps it
for modules, as they are executed once despite being potentially
imported multiple times...so it doesn't make sense for them to be
aware of the context of their caller, as they can have many callers.

But a new feature is added where resources can be addressed relative
to the currently running script.  That is done with TAG!

So if you DO a script like %/home/modules/script.r

    import <alpha.r>  ; will load from %/home/modules/
    import %beta.r   ; also loads from %/home/modules/

    change-dir %subdir/

    import <delta.r>  ; still loads from %/home/modules/
    import %gamma.r  ; loads from %/home/modules/subdir/

The previous function of TAG! as picking a module from the index
of modules via shorthand is being replaced by this.  But for interim
compatibility, a TAG! import with no dot in it will be treated like
the previous lookups.  Future uses should use another notation, and
currently the leading idea for that is a THE-WORD! type, like @json.