Make non-TAG! > and < illegal in PATH! and TUPLE!
This resolves an ambiguity in PATH!s and TUPLE!s with TAG!, favoring
the idea that dots and slashes should be legal in tags.

The result is allowing things like `<../foo/bar.txt>`, at the cost of
not being to call refinements on things like < (e.g. `</signed`) or
to pick members out of objects assigned to arrow words (`<-.field`)

Enforcement of the illegality of arrow words is done by giving the
symbol nodes themselves an annotation to say that they are "arrow
words", which is checked at the moment of tuple or path creation.