Change passing of `verb` to pass by REBSYM*
The verb passed to the generic dispatch (word like APPEND, or OPEN)
was being passed as a REBVAL* cell.

Passing as a REBSYM* is more narrow, so there's no way to fail on
if the REBVAL* is not a WORD!.  It simplifies calling for those who
don't want to allocate a whole cell.