Preliminary "PICK-POKE*", Begin phase out PD_Xxx
This begins reimagining pathing as something that is implemented via
a typical generic function dispatch, instead of using a special
frame variant (a "PVS").  The pickpoke method is based on a generic
model of recursion:

https://forum.rebol.info/t/1704

It also makes all SET-PATH! operations use POKE which sub-delegates to
the PICK-POKE* generic.  (The PICK mechanic currently involves pushing
function refinements to the stack, and hence the accumulation would
interfere with the generic conversion.)

While this is expected to have near-term detrimental impacts to
performance, it runs through more common mechanisms...where improvements
can be made systemic.  The PD_Xxx presented challenges to stackless
processing, and this is a better direction.