REB_0 as prep state for uninitialized cells
As a language, C provides some situations in which all-zero initialized
memory is easy to get.  For instance: global variables are always
initialized to all zero bytes when a process starts (which is generally
the way the OS provides it to avoid leakage of privileged information
from other processes).

So it would seem nice if one could use zero-initialized memory as the
"prepared space" for cells.  However, for safety Ren-C had been making
it so all slots that would be written into would have NODE_FLAG_NODE
and NODE_FLAG_CELL...and also wanted to have Prep_Cell() calls on
each location so that if it was a DEBUG_TRACK_EXTEND_CELLS build
then it could write the file and line number of the initialization.

This commit implements a compromise which still provides some safety
by making a header that's *all* zeros a distinguished state that is
considered safe to overwrite.  If it's not all zeros, then the usual
rules of requiring NODE_FLAG_NODE and NODE_FLAG_CELL apply.  This
provides a reasonable amount of safety while still permitting the
use of full zero initialized variables.