Force Init_XXX() functions to take RESET cells
This begins a concept that cells can only be written from an
emptied state.  The concept is that preventing the blind overwrite
of cell fields can ultimately lead to being able to reference count
the resources pointed to by cells.  That involves much more work,
but this is a necessary first step.

The idea actually creates some amount of clarity for following the
"hot potato" of when cells are live.  It makes Move_Cell() actually
useful as a distinction from Copy_Cell().  And it helps identify
places where things are being pointlessly double-initialized.  So
though it adds some amount of increased complexity, there's benefit.