Switch from trash for array terminator to free
Since the transition from using END markers to going by the tail of
arrays, marking the end of arrays with a terminator has been a
debug-build-only feature.  Keeping with history, within the active
portion of the array it hasn't stopped writes--only reads.

The type that was used previously was "trash", e.g. a BAD-WORD! with
a NULL label in the debug build.  However, reducing cells to a baseline
of REB_0 is being investigated as groundwork for being able to do
reference counting on series.

This changes from using trash to using the NODE_FLAG_FREE bit, which
was underutilized previously as it would block writes from happening.
Blocking writes can be done well enough with a "poison" type...but
for cases where you might like to stop reads but have the type be
different (e.g. this one) the flag can be used.