Simplify errors by accepting relative values
When specifiers were introduced, it was important not to lose the
specifier when moving values through the stack.  Errors tried to keep
this enforced, as with other code that built objects and blocks.

However, errors are not designed to carry code for execution purposes.
Many times the binding is not available, or difficult to get at.  It
may even be desirable to remove all binding from material passed to
errors to avoid binding "leakage".

With time, errors became more tolerant and could actually build
"unrelativized" values.  As the usage of these became more acceptable
it ceased to make sense to force the Error() constructor to only
take specific values that had gone through the unrelativization process.
Instead, it could be changed to do that unrelativization itself.

This makes that change, which simplifies a lot of code.