Update PROBE() to use SER_FLAVOR
The PROBE() code was an early Ren-C-ism that predates series flavors.
Flavors offer a parallel to cell types, and it makes Probe's job
easier.  This replaces the previous "guess series category or panic()"
behavior with switching on flavors and giving an appropriate diagnostic.

Output can be improved on an as-needed basis.