Attempt to fix unformatted cell issue
The Chainer_Dispatcher() was calling Process_Action() without going
through Eval() first, which was leading to it using SET_CELL_FLAG()
on a cell that hadn't gotten its NODE_FLAG_NODE or NODE_FLAG_CELL set.
This causes an assert in WRITABLE().

This refactors the code a little and uses raw bit flag setting instead
of going through the checked SET_CELL_FLAG().