Workaround longstanding MSVC optimizer warning
The unreachable code warning is useful, but there was a Catch-22 where
MSVC was noticing a function wouldn't return...only when it was
compiling the file it was in.  This meant a wrapper for this function
couldn't put a DEAD_END() annotation after the function because
sometimes the DEAD_END() trick itself would seem unreachable and
sometimes not.

This throws in a wrench the smart compiler can't guarantee doesn't
happen...by checking a runtime condition.  Hence it doesn't treat the
function like it has special knowedge when its source is available,
and a uniform wrapping can handle the warning consistently.