Fix and streamline POSIX CALL/SHELL
A problem was either introduced (or just exposed a missing behavior)
regarding how the POSIX code turned a block of input into a single
argv parameter in order to pass to the shell.

This tidies up the code and some of the hacks that existed prior to
the libRebol API.  Given the API's availability, several things that
were burdensome can be expressed more clearly.

This changes the behavior of TEXT! elements in blocks being passed
to the shell so that they won't preserve their outer quotes by
default.  The reasoning is explained in the code...doing anything else
would require some dialect design which overlaps with SHELL, and this
is intended to be a fairly low-level calling facility.