HELP fixes: Inherit description, specialize vs paths
Path-based specialization is a low-level mechanic which does not run
INHERIT-META to get HELP.  This means that defining CALL as :call*/wait
was not leading to any help.  For now, the answer is to use SPECIALIZE
instead.

Also, INHERIT-META was not inheriting a description if a new one was
not provided by REDESCRIBE.  This is usually probably an outdated
explanation (the parameter descriptions may be outdated too) but it's
perhaps better than nothing.