Remove _FILE_OFFSET_BITS from core system defines
R3-Alpha was more monolithic in terms of its build system, so things
like filesystem interaction was partly in the core--with some part
in a host abstraction.

Ren-C has moved toward being more modular, so that the core has no
real knowledge of a filesystem...and the filesystem extension takes the
place of a host abstraction layer.

Currently all seeking-file-IO is provided by a libuv extension.  The
compiler flags for _FILE_OFFSET_BITS are added or not based on that
extension, and only needed for those .c files (as the core does not
know or care about this flag).