Fix REQUEST-FILE/FILE refinement
It's possible to preload REQUEST-FILE with a filename.  But the code
doing that was broken, because it was getting the file size of a string
separately from the length...but taking the length of the original
string vs. the string content being the processed string.

The micro-optimization to save on a wcslen() call doesn't make sense.
This just gets the string once and gets the length.