Fix IMAGE! Incorrect Height Calculation
Although the code for IMAGE! is generally atrophied (and was based on
weak premises to begin with), it was used by some R3-Alpha programs
including the "USCII decoder".

This fixes a bug that was introduced in moving the image size position
out of the core.  The height was missing a division by a factor of
4 (e.g. 4 bytes per pixel).