Remove .xxx.yyy predicate form, simplify to ACTION!
Predicates were the idea that you could pass a function to something
like ANY or ALL that would determine the test used.

There was a shorthand for this taking advantage of the leading-BLANK!
form of TUPLE!:

    >> any .even? [1 4 7]
    == 4

While it was a nice idea, it introduced questions for more complex
functions about the position of the slot the predicate would occupy.

There has also been a change in thinking to where .xxx might be
taken for picking members out of the current object, as an alternative
syntax to self.xxx -- and since this hasn't been figured out it seems
premature to use this for predicates.

The idea is kept but by making /PREDICATE a normal refinement that
takes an ACTION!, for now.  A more succinct syntax may be added
later, but it's being put aside to work on higher priority features.