IMAGE! fixes for USCII
Because the IMAGE! code was sketchy, it was moved out of the core and
into an extension.  This freed up a builtin type-byte for other uses
and it was largely forgotten.  A simple image read/write test was
kept but nothing else.

Hence it was overlooked that the "image position" was being done on
the IMAGE! shared among all instances instead of on the image cell.
This changes that and also fixes the HEAD and TAIL reflectors.

(This is only being done to be able to get the USCII encoder/decoder
to run, for the purposes of testing other functions...it is not part
of an attempt to actually review or improve the IMAGE! concept.)