Adjust Redbol handling of INTEGER!+BINARY!
Appending an INTEGER! to a BINARY! in Rebol2 would append the digit
of the integer, not the byte value.  This is different in Red and in
R3-Alpha.

This decides against Rebol2 compatibility for the reason that it turns
out to be hard to figure out how to make and append a byte to a binary
if not via INTEGER!.