Back off RESET calls at initialization callsites
An experimental concept was tried with the new REB_0 mechanics for
reset cells, to have callites which overwrote existing cell material
be explicit about that by using RESET().

The idea was that it would offer a very minor speedup (avoiding a need
to test for REB_0 on the cell when reformatting it).  But beyond
giving that trivial benefit, it would not be so bad because it would
help make callsites clearer regarding where the "hot potato" of data
was...and encouraging cleanup of stray data.  This might increase
clarity and reduce bugs.

But it was actually causing bugs, as it's often useful to build a
new cell out of extracted material from the old cell.  So things like
this were causing undefined behavior:

    Init_Decimal(RESET(f->out), VAL_INTEGER(f->out));

Those are hard bugs to spot, and show why having mutating operations
as macros in function calls--either implicitly or explicitly--isn't
a great idea.

(Similar issues arose in early versions of TRACK(), and it was adjusted
to do all its modification out of band.)

So this scraps the idea and moves the logic for resetting inside the
Init_XXX functions as part of the header initialization process, not
done by a macro.

The experiment wasn't completely useless--as it did call attention to
some redundant initializations.