Make COLLECT trick work if rule fails
A trick to make PARSE3 return the result of a collect if it's the
only rule in the block only worked if the parse was successful.
Failure would not get to the point of changing the temporary variable
being assigned, and so it was getting an error due to being unset.

This initializes the temporary variable to NULL.