Fix unreadable ~trash~ in release build
The behavior of cells that are supposed to be write-only in the
release build is just to make them normal-looking ~trash~ bad words,
in the event that they should leak.  However, there had been some
issues with symbol load ordering.

Those issues were resolved by actually making the words list loading
distinct from the rest of boot.  However, it seems the trash cells
were being initialized with null symbols.

This wouldn't be noticed, since leaking trash shouldn't happen...but
this changes it to use the Canon(TRASH) and makes that available at
the right stage of booting.