Fix timezone compensation bug, linux time test
There was a sign error in Adjust_Date_Zone(), which was adding in
negative time zones instead of subtracting them.

This corrects that issue, and also on the Linux GCC build adds in a
test that compares the time and zone information with what the unix
`date --iso8601` command outputs.