Bigger Date Zone Adjustment
The displaying of dates and times was not consistent with system
times.  The compensation by changing a sign of a time zone fixed it,
but apparently disrupted the scanning of dates.  This makes the change
in what is hoped to be a more systemic way.

(This is more guesswork than is desirable.)