Move Complex Logic to SET/GET vs POKE/PICK
Because the central method for doing pathing was POKE* and PICK*, and
there was more parameterization possibility with them, it was at first
thought that the entry point to generic pathing and cached pathing
via processed "steps" would be done via POKE and PICK.

Looking more closely, it seems that GET and SET are better fits for
this concept, where their reduced arity is complemented by an added
return result to store the steps produced after processing the groups.
If this added result is not requested, then an error is raised if a
sequence containing groups is passed in.

This is a first draft of the idea, which switches the already
inefficient POKE call into a less efficient APPLY of SET with #
for the steps output call for SET-PATH! in the evaluator.  However,
it holds up well as a proof of concept...and allows the POKE and PICK
functions to move back to the simple notion of being 1-step of
dispatch only.