Remove pathing hook from "extension" datatypes
The concept of extending datatypes involved registering hooks for the
things that a datatype would do: a molding function, a path picking
function, a comparison function, TO and MAKE hooks, and a generic
catch-all "type hook".

With the changes to PICK* and POKE*, the idea is that users writing
ordinary functions could provide the functionality...not just people
writing C code.