Fix GROUP! at head of PATH! that aren't ACTION!
Historical Rebol/Red don't allow GROUP!s at the head of paths at all.
This means that when a GROUP! appears at the head of a path, the
rules can fit the new model as much as they want.

This makes them fit the model that PATH!s should only be used for
invoking actions.  (TUPLE!s are used for member picking, unless a
specific "Redbol Compatibility" mode is engaged.)