Tweak so stack overflow errors don't crash (as much?)
A test started triggering stack overflows in the code that sets up
a feed.  However the feed->value had been assigned to trash, and the
code that tried to clean up stack overflows was checking to see if
that was END or not.  (There actually seems to be code missing to
handle that case, which should be there...that needs a separate review)

This makes the code initialize instead to the END_CELL, which means
the cleanup code won't crash.