Start redesigning PATH! dispatch in evaluator
This begins to rethink the way that PATH! will work in the evaluator,
centering around the notion that it is used *only* to dispatch
functions.  The idea is that while tuple dispatch is based on the
PICK* and POKE* extensible hooks, path dispatch would really just
push refinements to the stack and do nothing else.

Since this is not compatible with the historical behavior of path
picking and poking, there will have to be some kind of compatibility
mode.  What will probably be done is that in the compatibility mode,
paths will be changed into TUPLE! and TUPLE! will specialize ACTION!
in its actions.