Add SYSTEM.OPTIONS.ACTION-ONLY-PATHS
This adds a workaround for old-style pathing which would have to be
enabled in order to use things like REDBOL.  What it does is adds a
means of turning ordinary picking into supporting specialization of
the action and returning it.

Besides not being desirable from a "confusing source" angle, letting
dots produce specialization in this way is inherently non-optimizable.
If path dispatch can be made non-extensible then it can be faster.