Use JOIN instead of MAKE-FILE in Bootstrap
MAKE-FILE was created as a cleaner answer to doing what pathing did
on FILE! historically.  It was used to replace those usages in the
bootstrap process.  However, JOIN suffices for these...as the extra
checking and flexibility that MAKE-FILE has is generally overkill
compared to how much of a hassle it creates to try and keep the
%make-file.r script working in both old and new interpreters.