Update %unzip.reb from dated workarounds, slash/dots
There were a couple of workarounds that weren't relevant anymore in
%unzip.reb...such as needing to use LIB.NOW instead of just NOW, and
mentioning the idea that LET variables holding dates couldn't be
updated.  Sea of words fixes the first, and a revised mechanism for
SET fixes the second.

There were some antiquated mentions of "ports" to translate dates
from, and old usages of slashes that needed to be dots.