Fix Sea of Words Module GC Bug
The concept in Sea of Words is that the global word table actually
holds a linked list of all the modules that word has a variable in.
Words can find the variable and actually cache a reference directly
to the variable's "node" (an 8-platform-pointer entity that holds the
cell).

What was happening was that the last thing that was done was that
all the contexts which were determined to still be alive through
references were getting their variables marked during a traversal
of the global word hash table.  But this had a weakness...because any
modules that became live during *that* walk would not then have their
own contents transitively marked (or only half marked, based on how
far through the word table enumeration it had gotten when the module
context became live).

In lieu of a terribly smart answer to the problem, this does something
relatively dumb...that is chosen for its other benefits.  What it
does is makes sure every time a node is marked that a counter is
increased.  The word table walk will continue cycling so long as
the count is different from what it was in the beginning.

Although this is not terribly efficient, it does offer the benefit
that the counter can be checked to make sure it balances with the
sweeping phase...so that everything is marked minimally as only once.