Remove level of indirection from symbol canons
This implements a variation on the trick that is used to make the
Lib(XXX) variable series, except making the symbol canon series nodes
actually reside in global memory.

It removes a level of indirection for Canon(XXX), and saves on the
allocation of the 800+ indirect pointers (about 6Kish on 64-bit).

Additionally, it makes it more "official" by actually making a
version of Make_Series() that's willing to take a node location to
write the bits into.  This gives things like being able to do the
allocation tracking in panic()...but also just generally to make sure
that anything which is done to series in general will also be done
for these series allocated outside of the memory pools.