Add simple ECHO function
With the arrival of Sea of Words, the concept is that you can use
words freely without worrying that every usage of a word is going to
allocate a variable.

    echo It opens new doors, like this one.

This makes an ECHO which can either be freeform like the above (a
traditional UNIX echo) or use a BLOCK!

    echo [Using a block means it's insensitive to newlines.]

This flips the burden to call out the insertions:

   print ["You have written" num-lines "lines of code."]

   echo [You have written @num-lines lines of code.]

In addition to being shorter, it has a clear advantage when written
in API calls, where you would have to escape strings (or use braces).