Make more efficient SET and GET functions
The first cut of SET and GET was written as their natives, so that
the evaluator had to recurse in order to do an assignment.  This
factors out the functionality in a way that it's easier to reuse.

This function offers the ability to either write or not write the
pick or poke STEPS that are taken during the set or get.  It also
changes the type of the steps when a block form to be THE-BLOCK!, as
the BLOCK! form may have a purpose, e.g. for symbols:

    [@]: default [1 + 2]