Simplify AND, OR, XOR behavior
The experimental AND/OR/XOR tried to fold in branching behavior as
ways to pass the left hand side to a branch on the right...as well as
attempting to make it so that GET-GROUP!s would evaluate regardless
of the short-circuiting.

This scales back those ideas to the much simpler idea of just handling
quoted GROUP!s, WORD!s, and TUPLE!s.