Pare back rebQ() and rebU() to arity-1 forms
An early design for rebQ() and rebU() was ambitious, in being variadic
and being able to span over ranges of splices...influencing the
quoting levels of values en masse.

The implementation was rather complex, and the functionality was not
used...with rebQ() and rebU() only being used in their arity-1 forms.
This removes the concept of the QUOTING_BYTE() from feeds and uses a
much simpler implementation.