Discontinue use of PROTO-PARSER for natives
The prototype parser was using actual LOAD logic to read in the
native prototypes in comment blocks.  That was a nice idea, but it
actually created a problem in that changes to what was legal in the
source format couldn't have those changes taken advantage of in the
native specifications, because they'd no longer load.

This is a somewhat rough change that drastically simplifies the
scanning, and processes the prototypes as text.  There's room for
improvement in terms of error detection and feedback, but for the
feature of being able to use new syntax in the spec blocks it's well
worth putting up with having more cryptic failure modes...especially
since it simplifies the source to where it's easier to follow to
catch the problems.