Some cleanup and comment fixes of %sys-series.h
This is a cleanup that started with an observation that the comment on
why LINK() and MISC() macros were necessary wasn't covering the fact
that they were only necessary due to C++ rules, as C allows type
punning of unions.

So this built out a better defense of the various reasons why the
LINK() and MISC() macros are good and useful, beyond just letting the
code compile correctly as C++.  That turned into a longer improvement
of some of the comments and organization of the file.

It also gets rid of the IS_SER_DYNAMIC() function, because encoding the
dynamism of a series in a SERIES_FLAG_XXX is now basically decided.
It looks better at the callsites and also provides the NOT_SERIES_FLAG
variation which reads more clearly.  (All things being equal, it
helps to know when what you're checking is determined by a flag.)