Streamline interface for INLINE, remove Cache_Predicate
There was residual carryover from the idea of making constructs
like `.not.even?` a shorthand way of expressing functions.  That idea
has been shelved for the moment.

This improves the INLINE functionality to allow you to inline either
a block, a quoted value, or a blank...giving callers some better
options for controlling an inlining site.