Get Rid of "Varlist Conservation"
Varlist conservation made it so that series that had been formatted
to be varlists would be reusable if an action that needed a similar
number of parameters came up.  This showed a noticeable performance
increase...however, methodology is improving so that similar gains
can probably be achieved without using a mechanism that adds such
unpredictability and complexity to the code.

In particular, it was in the way with difficult to interpret bugs, and
getting rid of it made more sense than fixing this particular
optimization.

Also improves CHAIN so that it uses the global inaccessible varlist
instead of creating a new one on every invocation.