Varlist uses Bonus for keysource, Link for patches
A concept being looked at is that objects and frames will be able to
be used directly as "virtual binding patches".  This requires them to
have the ability to point to a next patch.

For consistency, having this "NextPatch" use the same node slot as a
variable patch (e.g. a LET patch) is desirable.  Previously the
objects and frames were using the BONUS for their patch listings, but
this is not an option for a LET patch because the bonus lives where
the singular cell storage is (content/series bias).

This swaps it so that BONUS is used for something that applies to the
keylists...which applies to varlists specifically.  That opens up the
possibility of using their LINK for NextPatch, just like LETs do.