Canonize all inacccessible series to one node
The idea of being able to decay a series is to say that it has gone
away, but not freeing it because there may be live references to it
still.  So the GC has to wait until those references go away.

But since the only point is to have a value that doesn't crash, there's
not any need for more than one of these stub arrays that answers back
"inaccessible".

This changes the marking mechanics to actually pass pointers to the
nodes to mark so they can be changed and canonized to the global
inaccessible series.