Rewiring so patches use LINK for NextPatch
This tries to line up the NextPatch as using LINK() to match the
patch-related LINK() field in varlists.  It also removes the list
optimization that the varlist patch field was being used for previously
with the idea of using contexts as patches themselves...e.g. a frame
varlist could be a patch which links its fallback for lookup as
being to a module.

The goal of this is to be able to build a specifier that could do
a reconstruction of a binding environment--something like a scope.
This is really rough experimentation, just to see what might fall out
of the process.

Empirical tests suggested that the optimization lost in this commit
wasn't really providing much benefit.