Eliminate MONDEX Concept, Limit Quoting Byte
This puts in a more restrictive number of quoting levels, which makes
the full word index free for usage in REB_QUOTED instances.

It's a bit frustrating to imagine that a whole allocation of a series
node (8x uintptr_t size) is required just to get less than 8 bits freed
up in a cell to use for a quoting level.  But the idea is that it's
only supposed to happen rarely to get to 4 or more quote levels.

What this change is actually in preparation for is more creative use
of the binding bits than just for a simple index.