Change COPY of actions to not duplicate details
The original mode of making copies of actions was to copy their
parameter lists.  When actions were changed to use their "details"
array as identities, this switched to copying details.

However that is not safe.  Details arrays are black boxes, and might
contain (for instance) arbitrary C pointers that could be freed under
one action reference, and then incorrectly accessed through another
action reference that still had the copy.

This revisits the question with a new plan, that also simplifies the
mechanics of HIJACK.  This new idea is that an action looks in the
*archetype* slot for its details.  Hence the identity is usually--but
not always the same as the identity of the action in the first slot.

This has the somewhat interesting property that the ACT_ARCHETYPE()
is an action, just not necessarily the same one that was originally
referred to--in the case of a HIJACK.  This makes using an action's
archetype as a value act like the hijacked function.